requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pymongo==4.6.0
python-dateutil==2.8.2
flask==3.0.0
python-dotenv==1.0.0
